+++ src/main/java/uk/gov/companieshouse/orders/api/client/ApiClient.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.orders.api.service;
+package uk.gov.companieshouse.orders.api.client;
 
 import org.springframework.stereotype.Component;
 import uk.gov.companieshouse.api.InternalApiClient;
@@ -6,6 +6,7 @@ import uk.gov.companieshouse.sdk.manager.ApiSdkManager;
 
 @Component
 public class ApiClient {
+
     public InternalApiClient getInternalApiClient() {
         return ApiSdkManager.getPrivateSDK();
     }
