+++ src/main/java/uk/gov/companieshouse/orders/api/service/ApiClientService.java
@@ -4,6 +4,7 @@ import org.springframework.stereotype.Service;
 import uk.gov.companieshouse.api.handler.order.item.request.PrivateItemURIPattern;
 import uk.gov.companieshouse.api.handler.regex.URIValidator;
 import uk.gov.companieshouse.api.model.order.item.CertificateApi;
+import uk.gov.companieshouse.orders.api.client.ApiClient;
 import uk.gov.companieshouse.orders.api.mapper.ApiToCertificateMapper;
 import uk.gov.companieshouse.orders.api.model.Item;
